test(kernel): add resolveKernelLogArg seam + untagged level tests
Follow-up to the #401 review (finding 5, thread on kernel_test.go): the level-
resolution logic initKernelLogging feeds to kernel_init_logging was documented but
not tested. Extract the pure decision so it can be pinned by CI without cgo.

  - New untagged logging_level.go holds kernelLogLevel (moved out of cgo.go) and a
    new resolveKernelLogArg() (level string, useNULL bool): the NULL-on-
    DBSQL_KERNEL_DEBUG override vs the mapped driver level. Mirrors the existing
    errors_classify.go pattern (pure logic, no build tag) so its tests run under
    CGO_ENABLED=0. cgo.go's initKernelLogging is now a thin caller of it.
  - New untagged logging_level_test.go: TestKernelLogLevel (moved) + new
    TestResolveKernelLogArg, which pins the two invariants the PR justifies in prose
    — DBSQL_KERNEL_DEBUG (non-empty) yields useNULL=true so the kernel honors
    RUST_LOG, and otherwise the driver level is mapped in (incl. fatal→OFF). Empty
    value is treated as unset.
  - Drops the now-unused os import from cgo.go and zerolog from kernel_test.go.

Not covered (documented, cost/value flips): that the string is actually handed to
C.kernel_init_logging and the sync.Once first-call-wins — asserting those needs a
function-pointer seam over the cgo call plus resetting a package global, to test one
assignment + stdlib sync.Once.

Verified: CGO_ENABLED=0 build+vet+test green (both new tests run here); CGO_ENABLED=1
-tags databricks_kernel build+vet green, all four log tests pass.
